# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "BKRES API",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/device/create": {
      "post": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to create device",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "gateway_name": {
                  "example": "any"
                },
                "device_name": {
                  "example": "any"
                },
                "device_id": {
                  "example": "any"
                },
                "device_ip": {
                  "example": "any"
                },
                "count": {
                  "example": "any"
                },
                "mess_in_minute": {
                  "example": "any"
                },
                "is_block": {
                  "example": "any"
                },
                "time_interval": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/device/deletedevice/{API}": {
      "delete": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to delete device",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/device/updateDevice/{API}": {
      "put": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to update device",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "device_name": {
                  "example": "any"
                },
                "device_id": {
                  "example": "any"
                },
                "device_ip": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/device/getDevice/{API}": {
      "post": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to get device",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/device/getallDevice/{API}": {
      "post": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to get all device",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/data/sendData1": {
      "get": {
        "tags": [
          "Data"
        ],
        "description": "Endpoint to save data from device",
        "parameters": [
          {
            "name": "API",
            "in": "query",
            "type": "string"
          },
          {
            "name": "data",
            "in": "query",
            "type": "string"
          },
          {
            "name": "time",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/data/sendData2": {
      "post": {
        "tags": [
          "Data"
        ],
        "description": "Endpoint to save data from gateway",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/data/getdata/{API}": {
      "post": {
        "tags": [
          "Data"
        ],
        "description": "Endpoint to get data by API device",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}